Separate IrImageVector logic with Compose ImageVector
