+++ components/generator/imagevector/build.gradle.kts
@@ -9,7 +9,7 @@ dependencies {
     implementation(libs.kotlinpoet)
 
     testImplementation(projects.components.extensions)
-    testImplementation(projects.components.parser)
+    testImplementation(projects.components.parser.svgxml)
     testImplementation(libs.bundles.test)
     testRuntimeOnly(libs.junit.launcher)
 }
