+++ components/generator/imagevector/src/test/kotlin/io/github/composegears/valkyrie/generator/imagevector/ImageVectorWithPreviewTest.kt
@@ -4,7 +4,7 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import io.github.composegears.valkyrie.extensions.ResourceUtils.getResourcePath
 import io.github.composegears.valkyrie.generator.imagevector.common.toResourceText
-import io.github.composegears.valkyrie.parser.IconParser
+import io.github.composegears.valkyrie.parser.svgxml.IconParser
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
