+++ components/parser/ktfile/src/test/resources/AllPathParams.kt
