+++ components/parser/ktfile/src/test/resources/EmptyImageVector.kt
