+++ components/parser/ktfile/src/test/resources/EmptyPaths.kt
