+++ components/parser/ktfile/src/test/resources/IconWithGradient.kt
