+++ components/parser/ktfile/src/test/resources/IconWithGroup.kt
