+++ components/parser/ktfile/src/test/resources/LazyProperty.kt
