+++ components/parser/ktfile/src/test/resources/MaterialIcon.kt
