+++ components/parser/svgxml/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    alias(libs.plugins.kotlin.jvm)
+}
+
+dependencies {
+    implementation(projects.components.extensions)
+    api(projects.components.androidxCompose)
+
+    implementation(libs.android.build.tools)
+
+    testImplementation(libs.bundles.test)
+    testRuntimeOnly(libs.junit.launcher)
+}
