+++ components/psi/imagevector/src/main/kotlin/io/github/composegears/valkyrie/psi/imagevector/ImageVectorPsiParser.kt
@@ -1,6 +1,6 @@
 package io.github.composegears.valkyrie.psi.imagevector
 
-import androidx.compose.ui.graphics.vector.ImageVector
+import io.github.composegears.valkyrie.ir.IrImageVector
 import io.github.composegears.valkyrie.psi.imagevector.parser.MaterialImageVectorPsiParser
 import io.github.composegears.valkyrie.psi.imagevector.parser.RegularImageVectorPsiParser
 import org.jetbrains.kotlin.psi.KtFile
@@ -8,7 +8,7 @@ import org.jetbrains.kotlin.psi.KtImportList
 
 object ImageVectorPsiParser {
 
-    fun parseToImageVector(ktFile: KtFile): ImageVector? {
+    fun parseToIrImageVector(ktFile: KtFile): IrImageVector? {
         val isMaterial = ktFile.importList?.isMaterial() ?: return null
 
         return when {
