+++ idea-plugin/build.gradle.kts
@@ -12,9 +12,9 @@ dependencies {
     implementation(projects.components.extensions)
     implementation(projects.components.generator.iconpack)
     implementation(projects.components.generator.imagevector)
-    implementation(projects.components.parser)
+    implementation(projects.components.parser.ktfile)
+    implementation(projects.components.parser.svgxml)
     implementation(projects.components.psi.iconpack)
-    implementation(projects.components.psi.imagevector)
 
     compileOnly(compose.desktop.currentOs)
     implementation(compose.desktop.common)
