+++ idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/conversion/IconPackConversionViewModel.kt
@@ -1,12 +1,12 @@
 package io.github.composegears.valkyrie.ui.screen.mode.iconpack.conversion
 
 import com.composegears.tiamat.TiamatViewModel
+import io.github.composegears.valkyrie.extensions.isSvg
+import io.github.composegears.valkyrie.extensions.isXml
 import io.github.composegears.valkyrie.generator.imagevector.ImageVectorGenerator
 import io.github.composegears.valkyrie.generator.imagevector.ImageVectorGeneratorConfig
 import io.github.composegears.valkyrie.generator.imagevector.ImageVectorSpecOutput
-import io.github.composegears.valkyrie.parser.IconParser
-import io.github.composegears.valkyrie.parser.isSvg
-import io.github.composegears.valkyrie.parser.isXml
+import io.github.composegears.valkyrie.parser.svgxml.IconParser
 import io.github.composegears.valkyrie.processing.writter.FileWriter
 import io.github.composegears.valkyrie.settings.InMemorySettings
 import io.github.composegears.valkyrie.settings.ValkyriesSettings
