+++ settings.gradle.kts
@@ -52,7 +52,8 @@ include("components:generator:common")
 include("components:generator:iconpack")
 include("components:generator:imagevector")
 include("components:ir")
-include("components:parser")
+include("components:parser:ktfile")
+include("components:parser:svgxml")
 include("components:psi:iconpack")
 include("components:psi:imagevector")
 include("playground:app")
